requirements.txt
Flask
requests
matplotlib
pytest
python-dotenv 